+++ unityModule02/Assets/Script/IDamagable.cs
@@ -6,6 +6,6 @@ namespace Module02
 {
     public interface IDamagable
     {
-        public void Damaged(int damage);
+        public void Damaged(float damage);
     }
 }
